aur-fetch: remove diff functionality
This removes the following flags
* --log-dir
* --verbose
* --format

as well as the orderfile. This gives aur-fetch the distinct purpose of
retrieving and merging upstream changes from the AUR, and not serve
a second purpose of generating diffs for aur-sync or other tools.

Latter can be achieved with the extended --results output instead. See
also:

commit 85428c7077f28aab585dcd342c23df8d639dafaf
commit 655edbea647d5bf60fe9d96fa4660edebd8a4f78
commit c78c2613400afc32ca4588502b7a0697bb7b35f9